requirements.txt
flask
google-generativeai
pymupdf
python-dotenv
requests
beautifulsoup4
selenium